fix: correct totalWeight for zero-weight edges in Dijkstra path and int overflow in density
ShortestPathFinder.buildPath() was summing raw edge weights, but
findShortestByWeight() normalizes zero-weight edges to 1.0 during
traversal. This caused PathResult.getTotalWeight() to return 0 for
paths through zero-weight edges when the actual traversal cost was
positive. Added a normalizeZeroWeights parameter to buildPath() so
the weight normalization is consistent between traversal and result.

GraphStats.getDensity() computed v*(v-1) as int multiplication which
overflows for graphs with more than ~46K nodes (int max ~2.1B).
Cast to long to prevent silent overflow producing wrong density.

+++ Gvisual/src/gvisual/GraphStats.java
@@ -90,7 +90,7 @@ public class GraphStats {
         int v = getNodeCount();
         int e = getVisibleEdgeCount();
         if (v < 2) return 0.0;
-        return (2.0 * e) / (v * (v - 1));
+        return (2.0 * e) / ((long) v * (v - 1));
     }
 
     /**
